Add more conditions to the tests
