+++ boosters/node-beginner/meeting-3/src/server.js
@@ -1,11 +1,9 @@
 require("dotenv").config();
 const express = require("express");
 const routes = require("./api/routes");
-const bodyParser = require("body-parser");
 
 const app = express();
 
-app.use(bodyParser.json());
 app.get("/", (req, res) => {
   res.send("Hello World!");
 });
