+++ boosters/node-beginner/meeting-3/tests/04_exercise.test.js
@@ -1,8 +1,8 @@
 const request = require("supertest");
 const app = require("../src/server");
+const setup = require("./config/setup");
 const teardown = require("./config/teardown");
 const uuid = require("uuid/v4");
-const setup = require("./config/setup");
 
 const users = [
   { firstName: "Andrew", lastName: "Mayer", id: uuid() },
@@ -18,21 +18,19 @@ const repos = users.reduce(
   []
 );
 
-describe("Get Repo", () => {
-  beforeAll(async () => {
+describe("User Repos", () => {
+  beforeEach(async () => {
     await setup("Users", users)();
     return setup("Repos", repos)();
   });
+  afterEach(teardown);
+  xit("Should return all repos of a given userId", async () => {
+    res = await Promise.all(users.map(user => request(app).get(`/users/${user.id}/repos`)));
 
-  afterAll(teardown);
-
-  xit("Should get a specific repo", async () => {
-    const res = await Promise.all(repos.map(repo => request(app).get(`/repos/${repo.id}`)));
-    expect(res.length).toEqual(repos.length);
-    res.map((res, i) => {
+    res.map((res, index) => {
       expect(res.statusCode).toEqual(200);
-      expect(res.body.name).toEqual(repos[i].name);
-      expect(res.body.id).toEqual(repos[i].id);
+      expect(res.body.length).toEqual(2);
+      res.body.map(body => expect(body.userId).toEqual(users[index].id));
     });
   });
 });
